+++ src/converters/piral-ng/src/converter.ts
@@ -3,6 +3,7 @@ import type { ForeignComponent, BaseComponentProps } from 'piral-core';
 import { createExtension } from './extension';
 import { enqueue } from './queue';
 import { bootstrap, NgModuleInt } from './bootstrap';
+import { NgOptions } from '.';
 
 let next = ~~(Math.random() * 10000);
 
@@ -20,20 +21,20 @@ export interface NgConverterOptions {
   /**
    * Defines the module options to apply when bootstrapping a component.
    */
-  moduleOptions?: Omit<NgModule, 'boostrap'>;
+  moduleOptions?: Omit<NgModule, 'bootstrap'>;
 }
 
 export function createConverter(config: NgConverterOptions = {}) {
   const { selectId = () => `ng-${next++}`, moduleOptions = {}, selector = 'extension-component' } = config;
   const Extension = createExtension(selector);
-  const convert = <TProps extends BaseComponentProps>(component: any): ForeignComponent<TProps> => {
+  const convert = <TProps extends BaseComponentProps>(component: any, opts?: NgOptions): ForeignComponent<TProps> => {
     let result: Promise<void | NgModuleInt> = Promise.resolve();
     let active = true;
 
     return {
       mount(el, props, ctx) {
         const id = selectId();
-        result = enqueue(() => active && bootstrap(ctx, props, component, el, id, moduleOptions, Extension));
+        result = enqueue(() => active && bootstrap(ctx, props, component, el, id, moduleOptions, Extension, opts));
       },
       unmount() {
         active = false;